requirements.txt
Flask==2.1.2
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
flask-restx==0.5.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Werkzeug==2.1.2
python-dotenv==0.20.0
python-slugify==6.1.2
gunicorn==20.1.0
psycopg2-binary==2.9.3